# Conflicts:
#	packages/relay-runtime/store/RelayPublishQueue.js

+++ packages/relay-compiler/codegen/RelayFileWriter.js
@@ -10,7 +10,6 @@
 
 'use strict';
 
-const RelayFlowGenerator = require('../core/RelayFlowGenerator');
 const RelayParser = require('../core/RelayParser');
 const RelayValidator = require('../core/RelayValidator');
 
@@ -30,9 +29,12 @@ const {
 } = require('graphql-compiler');
 const {Map: ImmutableMap} = require('immutable');
 
-import type {ScalarTypeMapping} from '../core/RelayFlowTypeTransformers';
+import type {
+  FormatModule,
+  TypeGenerator,
+} from '../language/RelayLanguagePluginInterface';
+import type {ScalarTypeMapping} from '../language/javascript/RelayFlowTypeTransformers';
 import type {RelayCompilerTransforms} from './compileRelayArtifacts';
-import type {FormatModule} from './writeRelayGeneratedFile';
 import type {
   FileWriterInterface,
   Reporter,
@@ -57,13 +59,16 @@ export type WriterConfig = {
   formatModule: FormatModule,
   generateExtraFiles?: GenerateExtraFiles,
   inputFieldWhiteListForFlow: Array<string>,
-  outputDir?: string,
+  outputDir?: ?string,
+  generatedDirectories?: Array<string>,
   persistQuery?: (text: string) => Promise<string>,
   platform?: string,
   relayRuntimeModule?: string,
   schemaExtensions: Array<string>,
   noFutureProofEnums: boolean,
   useHaste: boolean,
+  extension: string,
+  typeGenerator: TypeGenerator,
   // Haste style module that exports flow types for GraphQL enums.
   // TODO(T22422153) support non-haste environments
   enumsHasteModule?: string,
@@ -71,6 +76,8 @@ export type WriterConfig = {
     GLOBAL_RULES?: Array<ValidationRule>,
     LOCAL_RULES?: Array<ValidationRule>,
   },
+  // EXPERIMENTAL: skips deleting extra files in the generated directories
+  experimental_noDeleteExtraFiles?: boolean,
 };
 
 class RelayFileWriter implements FileWriterInterface {
@@ -153,6 +160,10 @@ class RelayFileWriter implements FileWriterInterface {
         return codegenDir;
       };
 
+      for (const existingDirectory of this._config.generatedDirectories || []) {
+        addCodegenDir(existingDirectory);
+      }
+
       let configOutputDirectory;
       if (this._config.outputDir) {
         configOutputDirectory = addCodegenDir(this._config.outputDir);
@@ -213,7 +224,7 @@ class RelayFileWriter implements FileWriterInterface {
       };
 
       const transformedFlowContext = compilerContext.applyTransforms(
-        RelayFlowGenerator.flowTransforms,
+        this._config.typeGenerator.transforms,
         this._reporter,
       );
       const transformedQueryContext = compilerContext.applyTransforms(
@@ -266,20 +277,21 @@ class RelayFileWriter implements FileWriterInterface {
             const relayRuntimeModule =
               this._config.relayRuntimeModule || 'relay-runtime';
 
-            const flowNode = transformedFlowContext.get(node.name);
+            const typeNode = transformedFlowContext.get(node.name);
             invariant(
-              flowNode,
-              'RelayFileWriter: did not compile flow types for: %s',
+              typeNode,
+              'RelayFileWriter: did not compile types for: %s',
               node.name,
             );
 
-            const flowTypes = RelayFlowGenerator.generate(flowNode, {
+            const typeText = this._config.typeGenerator.generate(typeNode, {
               customScalars: this._config.customScalars,
               enumsHasteModule: this._config.enumsHasteModule,
               existingFragmentNames,
               inputFieldWhiteList: this._config.inputFieldWhiteListForFlow,
               relayRuntimeModule,
               useHaste: this._config.useHaste,
+              useSingleArtifactDirectory: !!this._config.outputDir,
               noFutureProofEnums: this._config.noFutureProofEnums,
             });
 
@@ -291,11 +303,12 @@ class RelayFileWriter implements FileWriterInterface {
               getGeneratedDirectory(node.name),
               node,
               formatModule,
-              flowTypes,
+              typeText,
               persistQuery,
               this._config.platform,
               relayRuntimeModule,
               sourceHash,
+              this._config.extension,
             );
           }),
         );
@@ -325,9 +338,11 @@ class RelayFileWriter implements FileWriterInterface {
         }
 
         // clean output directories
-        allOutputDirectories.forEach(dir => {
-          dir.deleteExtraFiles();
-        });
+        if (this._config.experimental_noDeleteExtraFiles !== true) {
+          allOutputDirectories.forEach(dir => {
+            dir.deleteExtraFiles();
+          });
+        }
         if (this._sourceControl && !this._onlyValidate) {
           await CodegenDirectory.sourceControlAddRemove(
             this._sourceControl,
@@ -338,7 +353,7 @@ class RelayFileWriter implements FileWriterInterface {
         let details;
         try {
           details = JSON.parse(error.message);
-        } catch (_) {}
+        } catch (_) {} // eslint-disable-line lint/no-unused-catch-bindings
         if (
           details &&
           details.name === 'GraphQL2Exception' &&
